Begin implementation of SimplePowerGogglesRenderer render background and gradient rectangle
